@@ -1,12 +1,12 @@
 {
   "name": "plaid",
   "productName": "Plaid",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "author": {
     "name": "Piotr Marczyński",
     "email": "piotmski@gmail.com"
   },
-  "description": "Get a better overview of your Jira work log",
+  "description": "Better manage your Jira work log",
   "keywords": [
     "jira",
     "work log",
@@ -52,34 +52,34 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^9.0.6",
-    "@angular/common": "^9.0.6",
-    "@angular/compiler": "^9.0.6",
-    "@angular/core": "^9.0.6",
-    "@angular/forms": "^9.0.6",
-    "@angular/platform-browser": "^9.0.6",
-    "@angular/platform-browser-dynamic": "^9.0.6",
-    "@angular/router": "^9.0.6",
+    "@angular/animations": "^9.1.1",
+    "@angular/common": "^9.1.1",
+    "@angular/compiler": "^9.1.1",
+    "@angular/core": "^9.1.1",
+    "@angular/forms": "^9.1.1",
+    "@angular/platform-browser": "^9.1.1",
+    "@angular/platform-browser-dynamic": "^9.1.1",
+    "@angular/router": "^9.1.1",
     "@openfonts/comfortaa_latin": "^1.44.3",
     "@openfonts/source-sans-pro_all": "^1.44.0",
     "electron-updater": "^4.2.5",
     "mousetrap": "^1.6.5",
     "ngx-electron": "~2.2.0",
-    "rxjs": "~6.5.4",
+    "rxjs": "~6.5.5",
     "tslib": "^1.11.1",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.900.6",
-    "@angular/cli": "^9.0.6",
-    "@angular/compiler-cli": "^9.0.6",
-    "@angular/language-service": "^9.0.6",
-    "@types/node": "^12.12.30",
-    "codelyzer": "^5.1.2",
-    "electron": "^8.1.1",
+    "@angular-devkit/build-angular": "^0.901.1",
+    "@angular/cli": "^9.1.1",
+    "@angular/compiler-cli": "^9.1.1",
+    "@angular/language-service": "^9.1.1",
+    "@types/node": "^12.12.35",
+    "codelyzer": "^5.2.2",
+    "electron": "^8.2.1",
     "electron-builder": "^22.4.1",
-    "ts-node": "~8.6.2",
-    "tslint": "~5.20.1",
-    "typescript": "~3.7.5"
+    "ts-node": "~8.8.2",
+    "tslint": "~6.1.1",
+    "typescript": "~3.8.3"
   }
 }
