@@ -7,14 +7,13 @@ import {mergeMap} from 'rxjs/operators';
 @Injectable({ providedIn: 'root' })
 export class AuthApi {
   private sessionUrl = '/rest/auth/1/session';
-  private userUrl = '/rest/api/2/user?username=';
 
   constructor(private http: HttpClient) { }
 
   getAuthenticatedUser$(): Observable<User> {
     return this.http.get<User>(this.sessionUrl).pipe(
       mergeMap<User, Observable<User>>(
-        sessionUser => this.http.get<User>(this.userUrl + sessionUser.name)
+        sessionUser => this.http.get<User>(sessionUser.self)
       )
     );
   }
