@@ -15,6 +15,8 @@ import {NgxElectronModule} from 'ngx-electron';
 import {AuthStatusComponent} from './components/auth-status/auth-status.component';
 import {ZoomControlsComponent} from './components/zoom-controls/zoom-controls.component';
 import {CurrentTimeMarkerComponent} from './components/current-time-marker/current-time-marker.component';
+import {WorklogEditorComponent} from './components/worklog-editor/worklog-editor.component';
+import {DatePickerCloudComponent} from './components/date-picker-cloud/date-picker-cloud.component';
 
 @NgModule({
   declarations: [
@@ -23,12 +25,14 @@ import {CurrentTimeMarkerComponent} from './components/current-time-marker/curre
     WorklogPanelComponent,
     ConnectionIssueResolverComponent,
     DateRangePickerComponent,
+    DatePickerCloudComponent,
     GridBackgroundComponent,
     RefreshButtonComponent,
     ExtHrefDirective,
     AuthStatusComponent,
     ZoomControlsComponent,
-    CurrentTimeMarkerComponent
+    CurrentTimeMarkerComponent,
+    WorklogEditorComponent
   ],
   imports: [
     BrowserModule,
